+++ tests/test/02_shmem/procinfo_01.c
@@ -189,6 +189,36 @@ int main( int argc, char **argv ) {
                 assert( shmem_procinfo_ext__recover_stolen_cpus(222) == DLB_NOUPDT );
             }
 
+            // test stealing all CPUs
+            {
+                // subprocess 1 gets all CPUs
+                mu_parse_mask("0-3", &mask);
+                assert( shmem_procinfo__setprocessmask(111, &mask, flags) == DLB_SUCCESS );
+                // check that subprocess 2 has empty mask
+                assert( shmem_procinfo__polldrom(222, NULL, &mask) == DLB_SUCCESS );
+                assert( CPU_COUNT(&mask) == 0 );
+                // check that subprocess 1 has CPUs 0-3
+                assert( shmem_procinfo__polldrom(111, NULL, &mask) == DLB_SUCCESS );
+                assert( CPU_COUNT(&mask) == 4
+                        && CPU_ISSET(0, &mask) && CPU_ISSET(1, &mask)
+                        && CPU_ISSET(2, &mask) && CPU_ISSET(3, &mask) );
+
+                // set original mask to subprocess 1 (WITH return_stolen)
+                flags |= DLB_RETURN_STOLEN;
+                assert( shmem_procinfo__setprocessmask(111, &p1_mask, flags) == DLB_SUCCESS );
+                // check that subprocess 2 has CPUs 2-3
+                assert( shmem_procinfo__polldrom(222, NULL, &mask) == DLB_SUCCESS );
+                assert( CPU_COUNT(&mask) == 2
+                        && CPU_ISSET(2, &mask) && CPU_ISSET(3, &mask) );
+                // check that subprocess 1 has CPUs 0-1
+                assert( shmem_procinfo__polldrom(111, NULL, &mask) == DLB_SUCCESS );
+                assert( CPU_COUNT(&mask) == 2
+                        && CPU_ISSET(0, &mask) && CPU_ISSET(1, &mask) );
+
+                // check that there are not stolen CPUs
+                assert( shmem_procinfo_ext__recover_stolen_cpus(222) == DLB_NOUPDT );
+            }
+
             assert( shmem_procinfo__finalize(111, false, SHMEM_KEY) == DLB_SUCCESS );
             assert( shmem_procinfo__finalize(222, false, SHMEM_KEY) == DLB_SUCCESS );
         }